Main page fetch revalidation added

+++ src/utils/sessionStorage.ts
@@ -1,15 +1,26 @@
+"use client"
+
 export function setToSessionStorage(key: string, value: string) {
-  sessionStorage.setItem(key, value);
+  if (typeof window !== 'undefined') {
+
+    return sessionStorage.setItem(key, value);
+  }
 }
 
 export function getFromSessionStorage(key: string) {
-  return JSON.parse(sessionStorage.getItem(key) as string);
+  if (typeof window !== 'undefined') {
+    return JSON.parse(sessionStorage.getItem(key) as string);
+  }
 }
 
 export function removeFromSessionStorage(key: string) {
-  sessionStorage.removeItem(key);
+  if (typeof window !== 'undefined') {
+    return sessionStorage.removeItem(key);
+  }
 }
 
 export function clearSessionStorage() {
-  sessionStorage.clear();
+  if (typeof window !== 'undefined') {
+    return sessionStorage.clear();
+  }
 }
